// application/dto/main/types.gen.ts
export type Task = {
    /**
     * Title
     */
    title?: string | null;
    /**
     * Description
     */
    description?: string | null;
    /**
     * User Id
     */
    user_id?: string | null;
    /**
     * Is Completed
     */
    is_completed?: boolean | null;
    /**
     * Is Accepted
     */
    is_accepted?: boolean | null;
    /**
     * Priority
     */
    priority?: number | null;
    /**
     * Deadline
     */
    deadline?: string | null;
    /**
     * Duration
     */
    duration?: number | null;
    /**
     * Additional Tag
     */
    additional_tag?: string | null;
    /**
     * Event Type Id
     */
    event_type_id?: number | null;
    /**
     * Id
     */
    id: number;
    /**
     * Is Active
     */
    is_active: boolean;
    /**
     * Updated On
     */
    updated_on: string;
    /**
     * Created On
     */
    created_on: string;
};